Add print_grid() and list of ideas
I'm thinking that this project will be a weekend project.
May I work on it for a couple of hours each weekend. Or, during
the week when I get bored, I might mess around and do some work
on this. I genuinely enjoy writing this code and I think with a lil
reseach, I can make this a solid project.