Fix algorithm for checking quadrants
The last version of code that checked quadrants didn't iterate through
them properly. This error was caught using unit tests (new BFFs).

Now, the code uses a pair of for loops that each run 3 times to give the
beginning row and beginning column for each quadrant. I believe that
this commit is fully playable (albeit terribly).