Correct guess checking logic
The old logic would use the XXX_clear functions to see if a guess was
godo and the only time these functions would return True was when there
were 80 spots filled in correctly and you correctly guessed the 81st
element. The new functions (check_XXX) just makes sure that guess isn't
in the row, col, or quadrant of the guess.

The XXX_clear functions are still useful in game_over() but I'm sure
there is some way to chop down this code much further.