Cut fat from gen_quad, add tests
The generate() still doesn't work and I don't know why. I'm hoping to
get to the bottom of this through writing tests and making sure that all
componenets (methods) work as I expect them to

In the old version of gen_quad, I relied on the fact that two np.NaNs
are unique which meant that type conversion had to be done. Now, the
number of uniques of a quadrant get compared with the number of unique
nonzero values of a quad. If those are equal, generate the quad, if
they're unequal, generate another one.