Backtracking now works! Sorta!
After lots of blood and tears, backtracking now almost works
completely. Issue is, it doesn't work with really hard puzzles. I have a
feeling that the reason for this is that it doesn't keep track of what
trees we've gone down so it goes down the same tree multiple times,
obsviously, this is a waste of memory and very inefficient. I'm not 100%
sure if this is the case though.