Start JS implmenetation.
Right now, the code renders a 9x9 double array to the broswer and it
resembles a sudoku somewhat. Now we can just operate on the grid and
then those operations should be reflected in the grid. This may need
some sort of react or something.

* Still need to get actual sudoku logic in
* Still need to implement the visual version of a backtracking algo
* Need to figure out some way to get user input

I learned that I'm not the biggest fan of making UIs/visuals