Start load/save game functionality
I'm not going to do the whole hash in URL thing.
There's going to be an option for a use to get the 81
character string and then they will be able to load a game
or generate a game when they first arrive. For now, the only
option will be to load a game